refactor: unify module metadata
