--- src/modules.js
@@ -1,8 +1,0 @@
-export const modules = [
-  {
-    id: 'module1',
-    title: "Module 1: You're in! First steps to data analysis",
-    csv: '/assets/quiz_questions.csv'
-  }
-  // Additional modules can be added here
-];
